Make linter aware of `enableGlobDot` config option.
