+++ packages/cspell/samples/hidden-test/.hidden/.text.md
@@ -1,0 +1,1 @@
+# Nested Hidden file.
