+++ packages/cspell/samples/hidden-test/.hidden/hidden.md
