+++ packages/cspell/src/lint/index.ts
@@ -1,0 +1,1 @@
+export { runLint } from './lint';
